--- src/main/java/mate/academy/bookapp/service/BookServiceImpl.java
@@ -1,95 +1,0 @@
-package mate.academy.bookapp.service;
-
-import jakarta.transaction.Transactional;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import mate.academy.bookapp.dto.book.BookDto;
-import mate.academy.bookapp.dto.book.BookDtoWithoutCategoryIds;
-import mate.academy.bookapp.dto.book.BookSearchParametersDto;
-import mate.academy.bookapp.dto.book.CreateBookRequestDto;
-import mate.academy.bookapp.exceptions.EntityNotFoundException;
-import mate.academy.bookapp.mapper.BookMapper;
-import mate.academy.bookapp.model.Book;
-import mate.academy.bookapp.model.Category;
-import mate.academy.bookapp.repository.book.BookRepository;
-import mate.academy.bookapp.repository.book.BookSpecificationBuilder;
-import mate.academy.bookapp.repository.category.CategoryRepository;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.stereotype.Service;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class BookServiceImpl implements BookService {
-    private final BookRepository bookRepository;
-    private final BookMapper bookMapper;
-    private final BookSpecificationBuilder bookSpecificationBuilder;
-    private final CategoryRepository categoryRepository;
-
-    @Override
-    public BookDto save(CreateBookRequestDto requestDto) {
-        Book book = bookMapper.toModel(requestDto);
-
-        Set<Category> categories = mapCategoryIdsToCategories(
-                requestDto.getCategoryIds());
-        book.setCategories(categories);
-
-        return bookMapper.toDto(bookRepository.save(book));
-    }
-
-    @Override
-    public BookDto getById(Long id) {
-        Book book = bookRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Book with id " + id + " not found")
-        );
-        return bookMapper.toDto(book);
-    }
-
-    @Override
-    public Page<BookDto> findAll(Pageable pageable) {
-        return bookRepository.findAll(pageable)
-                .map(bookMapper::toDto);
-    }
-
-    @Override
-    public void deleteById(Long id) {
-        bookRepository.deleteById(id);
-    }
-
-    @Override
-    public BookDto updateBook(Long id, CreateBookRequestDto requestDto) {
-        Book book = bookRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Book with id " + id + " not found")
-        );
-        bookMapper.updateBook(requestDto, book);
-        Set<Category> categories = mapCategoryIdsToCategories(
-                requestDto.getCategoryIds());
-        book.setCategories(categories);
-        return bookMapper.toDto(bookRepository.save(book));
-    }
-
-    @Override
-    public Page<BookDto> search(BookSearchParametersDto params, Pageable pageable) {
-        Specification<Book> bookSpecification = bookSpecificationBuilder.build(params);
-        return bookRepository.findAll(bookSpecification, pageable)
-                .map(bookMapper::toDto);
-    }
-
-    @Override
-    public Page<BookDtoWithoutCategoryIds> getAllByCategoryId(Long categoryId, Pageable pageable) {
-        return bookRepository.findAllByCategoryId(categoryId, pageable)
-                .map(bookMapper::toDtoWithoutCategories);
-    }
-
-    public Set<Category> mapCategoryIdsToCategories(List<Long> categoryIds) {
-        return categoryIds.stream()
-                .map(categoryId -> categoryRepository.findById(categoryId)
-                        .orElseThrow(() -> new EntityNotFoundException("Category not found by id: "
-                                + categoryId)))
-                .collect(Collectors.toSet());
-    }
-}
